new checks

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -9,12 +9,12 @@
 int main(void)
 {
 	int i;
-	int n1 = 0, n2 = 1, sum;
+	long n1 = 0, n2 = 1, sum;
 
 	for (i = 0; i < 50; i++)
 	{
 		sum = n1 + n2;
-		printf("%d", sum);
+		printf("%ld", sum);
 
 		n1 = n2;
 		n2 = sum;
